separarted results in another file
